- Changed `github_release` task to work with an existing tag/ref if   already in the repo and sleep for 3 seconds for Github to catch up after creating the tag - Removed unneeded 5 second sleep in release_beta flow

+++ cumulusci/tasks/github.py
@@ -1,3 +1,5 @@
+import time
+
 from datetime import datetime
 from github3 import login
 
@@ -96,32 +98,38 @@ class CreateRelease(BaseGithubTask):
         version = self.options['version']
         self.tag_name = self.project_config.get_tag_for_version(version)
 
-        # Create the annotated tag
-        tag = repo.create_tag(
-            tag = self.tag_name,
-            message = 'Release of version {}'.format(version),
-            sha = commit,
-            obj_type = 'commit',
-            tagger = {
-                'name': self.github_config.username,
-                'email': self.github_config.email,
-                'date': '{}Z'.format(datetime.now().isoformat()),
-            },
-            lightweight = False,
-        )
-
-        # Get the ref created from the previous call that for some reason creates
-        # a ref to the commit sha rather than the tag sha.  Delete the ref so we
-        # can create the right one.  FIXME: Is this a bug in github3.py?
         ref = repo.ref('tags/{}'.format(self.tag_name))
-        if ref:
-            ref.delete()
 
-        # Create the ref linking to the tag
-        ref = repo.create_ref(
-            ref = 'refs/tags/{}'.format(self.tag_name),
-            sha = tag.sha,
-        )
+        if not ref:
+            # Create the annotated tag
+            tag = repo.create_tag(
+                tag = self.tag_name,
+                message = 'Release of version {}'.format(version),
+                sha = commit,
+                obj_type = 'commit',
+                tagger = {
+                    'name': self.github_config.username,
+                    'email': self.github_config.email,
+                    'date': '{}Z'.format(datetime.now().isoformat()),
+                },
+                lightweight = False,
+            )
+    
+            # Get the ref created from the previous call that for some reason creates
+            # a ref to the commit sha rather than the tag sha.  Delete the ref so we
+            # can create the right one.  FIXME: Is this a bug in github3.py?
+            ref = repo.ref('tags/{}'.format(self.tag_name))
+            if ref:
+                ref.delete()
+    
+            # Create the ref linking to the tag
+            ref = repo.create_ref(
+                ref = 'refs/tags/{}'.format(self.tag_name),
+                sha = tag.sha,
+            )
+
+            # Sleep for Github to catch up with the fact that the tag actually exists!
+            time.sleep(3)
 
         prerelease = 'Beta' in version
